+++ utils.h
@@ -2,4 +2,4 @@
 #include <Arduino.h>
 
 float getTorque(float rpm);
-uint8_t getCoolantTemp(uint8_t AA);
+int16_t getCoolantTemp(uint8_t AA);
